backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
torch==2.7.1
transformers==4.54.1
sentence-transformers==5.0.0
faiss-cpu==1.11.0.post1
python-dotenv==1.1.1
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
matplotlib>=3.8.0
numpy>=1.24.0